// src/api/generated/nvdApiTypes.ts
export interface JSONSchemaForNVDVulnerabilityDataAPIVersion222 {
	resultsPerPage: number;
	startIndex: number;
	totalResults: number;
	format: string;
	version: string;
	timestamp: string;
	/**
	 * NVD feed array of CVE
	 */
	vulnerabilities: DefCveItem[];
}
export interface DefCveItem {
	cve: CveItem;
}
export interface CveItem {
	id: CveId;
	sourceIdentifier?: string;
	vulnStatus?: string;
	published: string;
	lastModified: string;
	evaluatorComment?: string;
	evaluatorSolution?: string;
	evaluatorImpact?: string;
	cisaExploitAdd?: string;
	cisaActionDue?: string;
	cisaRequiredAction?: string;
	cisaVulnerabilityName?: string;
	cveTags?: {
		/**
		 * The email address or UUID of the source that contributed the information
		 */
		sourceIdentifier?: string;
		tags?: (
			| 'unsupported-when-assigned'
			| 'exclusively-hosted-service'
			| 'disputed'
		)[];
		[k: string]: unknown;
	}[];
	/**
	 * @minItems 1
	 */
	descriptions: [LangString, ...LangString[]];
	references: Reference[];
	/**
	 * Metric scores for a vulnerability as found on NVD.
	 */
	metrics?: {
		/**
		 * CVSS V4.0 score.
		 */
		cvssMetricV40?: CvssV40[];
		/**
		 * CVSS V3.1 score.
		 */
		cvssMetricV31?: CvssV31[];
		/**
		 * CVSS V3.0 score.
		 */
		cvssMetricV30?: CvssV30[];
		/**
		 * CVSS V2.0 score.
		 */
		cvssMetricV2?: CvssV2[];
		[k: string]: unknown;
	};
	weaknesses?: Weakness[];
	configurations?: Config[];
	vendorComments?: VendorComment[];
	[k: string]: unknown;
}